moves safe quotes to utils
